# docs/api-contracts/openapi-v1-baseline.yaml
...
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Rotate a BFF-held refresh token
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  schemas:
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
      - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/AuthUser'
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        status:
          type: string
...